# Conflicts:
#	src/com/cpsc304/JDBC/CustomerDBC.java
#	src/com/cpsc304/JDBC/LoginDBC.java
#	src/com/cpsc304/JDBC/UserDBC.java
#	src/com/cpsc304/Test/JDBC/CustomerDBCTest.java

+++ src/com/cpsc304/JDBC/UserDBC.java
@@ -6,7 +6,6 @@ import com.cpsc304.model.Restaurant;
 import com.cpsc304.model.User;
 
 import java.math.BigDecimal;
-import java.math.BigInteger;
 import java.sql.*;
 import java.util.List;
 import java.util.Set;
@@ -23,7 +22,7 @@ public abstract class UserDBC {
         pstmt = con.prepareStatement(sqlString);
         pstmt.setString(1, user.getName());
         pstmt.setString(2, user.getPassword());
-        pstmt.setLong(3, Long.parseLong(user.getPhoneNum()));
+        pstmt.setInt(3, Integer.parseInt(user.getPhoneNum()));
     }
 
     public static User getUser(String userID) throws SQLException {
